requirements.txt
# Use command 'pip freeze' to list all installed packages.
black==24.2.0
cachetools==5.3.2
click==8.1.7
google-auth==2.28.0
google-auth-oauthlib==1.2.0
gspread==6.0.2
oauthlib==3.2.2
pathspec==0.12.1
prettytable==3.10.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests-oauthlib==1.3.1
rsa==4.9
StrEnum==0.4.15